refactor(test): migrate skills.ts helpers to three-tier test architecture
Move 4 seeders (seedTestSkill, reseedSkills, bindCustomSkillToAgent,
createTestZeroSkill) to db-test-seeders/skills.ts and 2 assertions
(findTestSkillByUrl, getAgentCustomSkills) to db-test-assertions/skills.ts.
Replace api-test-helpers/skills.ts with pure re-exports.

Closes #9361
